Rename Page as PageProvider and move it with section into /components
