+++ docs/src/section/Section.mdx
@@ -6,7 +6,7 @@ menu: Sections
 
 import { Playground, Props } from 'docz'
 import {
-  Page,
+  PageProvider,
   Section,
   StickySectionWithActiveInfo,
 } from '@react-scrolly/core'
@@ -16,7 +16,7 @@ import { sectionStyle } from './styles'
 # Section
 <Playground>
   <div style={{ width: '100%', background: 'linear-gradient(135deg, #f5f7fa 0%, #c3cfe2 100%)' }}>
-    <Page>
+    <PageProvider>
       <StickySectionWithActiveInfo
         renderNonSticky={
           <>
@@ -50,7 +50,7 @@ import { sectionStyle } from './styles'
           </div>
         )}
       </StickySectionWithActiveInfo>
-    </Page>
+    </PageProvider>
   </div>
 </Playground>
